Most spent category data display in insights screen

+++ app/src/main/java/com/expensey/data/models/ExpenseSummary.kt
@@ -5,7 +5,7 @@ import java.math.BigDecimal
 data class ExpenseSummary(
     val categoryName: String,
     val categoryCount: Int,
-    val amountSpent: BigDecimal,
+    val amountSpent: Double,
     val categoryId: Int
 )
